+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add "(adjective) in (adjective) way" translation
 - Handle complicated phrase as error
 - Rearrange output to make adjective phrase appear first
+- Add basic "la" translation
 
 ## 0.0.2